Select's options reimplementation
