+++ lib/test/Select.test.js
@@ -323,7 +323,7 @@ describe("Select component tests", () => {
     expect(getAllByRole("option")[2].getAttribute("aria-selected")).toBe("true");
     expect(submitInput.value).toBe("3");
   });
-  test("Non-Grouped Options - Optional renders an empty first option with the placeholder as its label", () => {
+  test("Non-Grouped Options - Optional renders an empty first option (selected by default) with the placeholder as its label", () => {
     const onChange = jest.fn();
     const { getByRole, getAllByRole, getAllByText } = render(
       <DxcSelect
